responsive design work

+++ app/src/main/java/com/example/magsatpro/data/remote/AuthAPI.kt
@@ -8,7 +8,7 @@ import retrofit2.http.Query
 interface AuthAPI {
     @GET("android/login")
     suspend fun login(
-        @Query("code") code: Int,
+        @Query("code") code: Long,
         @Query("mac") mac: String,
         @Query("model") model : String,
